Move head.jsx and badge.jsx to components/
