+++ server/pages/components/head.jsx
